requirements.txt
Pillow==9.5.0
Kivy==2.1.0
